Спринт № 12, сделал отображение истории поиска
