+++ app/src/main/res/layout/activity_search.xml
@@ -75,8 +75,8 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:gravity="center_horizontal"
-                android:layout_marginTop="16dp"
-                android:layout_marginBottom="24dp"
+                android:layout_marginTop="36dp"
+                android:layout_marginBottom="20dp"
                 android:layout_marginHorizontal="24dp"
                 android:textColor="?attr/colorOnSecondary"
                 android:fontFamily="@font/ys_display_medium"
@@ -98,6 +98,8 @@
                 android:layout_height="36dp"
                 app:cornerRadius="18dp"
                 android:layout_gravity="center_horizontal"
+                android:layout_marginTop="24dp"
+                android:layout_marginBottom="20dp"
                 android:backgroundTint="?attr/colorOnSecondary"
                 android:textColor="?attr/colorOnTertiary"
                 android:textAllCaps="false"
